# Add AVX512 version of compare256