--- a/arch/x86/Makefile.in
+++ b/arch/x86/Makefile.in
@@ -36,6 +36,7 @@ all: \
 	chunkset_ssse3.o chunkset_ssse3.lo \
 	chorba_sse2.o chorba_sse2.lo \
 	compare256_avx2.o compare256_avx2.lo \
+	compare256_avx512.o compare256_avx512.lo \
 	compare256_sse2.o compare256_sse2.lo \
 	crc32_pclmulqdq.o crc32_pclmulqdq.lo \
 	crc32_vpclmulqdq.o crc32_vpclmulqdq.lo \
@@ -83,6 +84,12 @@ compare256_avx2.o:
 
 compare256_avx2.lo:
 	$(CC) $(SFLAGS) $(AVX2FLAG) $(NOLTOFLAG) -DPIC $(INCLUDES) -c -o $@ $(SRCDIR)/compare256_avx2.c
+
+compare256_avx512.o:
+	$(CC) $(CFLAGS) $(AVX512FLAG) $(NOLTOFLAG) $(INCLUDES) -c -o $@ $(SRCDIR)/compare256_avx512.c
+
+compare256_avx512.lo:
+	$(CC) $(SFLAGS) $(AVX512FLAG) $(NOLTOFLAG) -DPIC $(INCLUDES) -c -o $@ $(SRCDIR)/compare256_avx512.c
 
 compare256_sse2.o:
 	$(CC) $(CFLAGS) $(SSE2FLAG) $(NOLTOFLAG) $(INCLUDES) -c -o $@ $(SRCDIR)/compare256_sse2.c