--- a/arch/x86/x86_functions.h
+++ b/arch/x86/x86_functions.h
@@ -60,6 +60,11 @@ uint32_t adler32_fold_copy_avx512(uint32_t adler, uint8_t *dst, const uint8_t *s
 uint32_t chunksize_avx512(void);
 uint8_t* chunkmemset_safe_avx512(uint8_t *out, uint8_t *from, unsigned len, unsigned left);
 void inflate_fast_avx512(PREFIX3(stream)* strm, uint32_t start);
+#  ifdef HAVE_BUILTIN_CTZLL
+    uint32_t compare256_avx512(const uint8_t *src0, const uint8_t *src1);
+    uint32_t longest_match_avx512(deflate_state *const s, Pos cur_match);
+    uint32_t longest_match_slow_avx512(deflate_state *const s, Pos cur_match);
+#  endif
 #endif
 #ifdef X86_AVX512VNNI
 uint32_t adler32_avx512_vnni(uint32_t adler, const uint8_t *buf, size_t len);
@@ -169,6 +174,14 @@ uint32_t crc32_vpclmulqdq(uint32_t crc32, const uint8_t *buf, size_t len);
 #    define native_chunksize chunksize_avx512
 #    undef native_inflate_fast
 #    define native_inflate_fast inflate_fast_avx512
+#    ifdef HAVE_BUILTIN_CTZLL
+#      undef native_compare256
+#      define native_compare256 compare256_avx512
+#      undef native_longest_match
+#      define native_longest_match longest_match_avx512
+#      undef native_longest_match_slow
+#      define native_longest_match_slow longest_match_slow_avx512
+#    endif
 // X86 - AVX512 (VNNI)
 #    if defined(X86_AVX512VNNI) && defined(__AVX512VNNI__)
 #      undef native_adler32