Implements AVX512 variant of compare256.

Unlike AVX2, AVX512 does not need to make the mask separately, and instead can generate the mask directly from the comparison, saving us a tiny bit of time.

Using 64-byte compares from the start turned out to be a little slower on normal files because most matches are short or even not matches at all. Therefore I implemented this using two rounds of 32-byte compares at the start, before going up to 64byte compares after the first 64bytes have matched.

Before:
```
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     54.185%      0.078/0.078/0.079/0.000        0.030/0.031/0.031/0.000        8,526,745
 2     43.871%      0.128/0.129/0.129/0.000        0.030/0.030/0.030/0.000        6,903,702
 3     42.390%      0.158/0.158/0.158/0.000        0.029/0.029/0.029/0.000        6,670,664
 4     41.644%      0.181/0.181/0.182/0.000        0.028/0.028/0.028/0.000        6,553,205
 5     41.215%      0.196/0.197/0.197/0.000        0.028/0.028/0.028/0.000        6,485,659
 6     41.032%      0.243/0.244/0.244/0.000        0.028/0.028/0.028/0.000        6,456,912
 7     40.778%      0.337/0.338/0.338/0.000        0.028/0.028/0.028/0.000        6,416,941
 8     40.704%      0.445/0.446/0.446/0.000        0.028/0.028/0.028/0.000        6,405,249
 9     40.409%      0.528/0.528/0.529/0.000        0.027/0.027/0.027/0.000        6,358,951

 avg1  42.914%                        0.255                          0.029
 tot                                 68.956                          7.705       60,778,028

   text    data     bss     dec     hex filename
 170024    1344       8  171376   29d70 libz-ng.so.2
```

After:
```
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     54.185%      0.079/0.079/0.079/0.000        0.031/0.031/0.031/0.000        8,526,745
 2     43.871%      0.127/0.128/0.128/0.000        0.030/0.030/0.030/0.000        6,903,702
 3     42.390%      0.156/0.156/0.157/0.000        0.029/0.029/0.029/0.000        6,670,664
 4     41.644%      0.178/0.179/0.180/0.000        0.028/0.028/0.028/0.000        6,553,205
 5     41.215%      0.193/0.194/0.194/0.000        0.028/0.028/0.028/0.000        6,485,659
 6     41.032%      0.240/0.240/0.240/0.000        0.028/0.028/0.028/0.000        6,456,912
 7     40.778%      0.335/0.335/0.335/0.000        0.028/0.028/0.028/0.000        6,416,941
 8     40.704%      0.442/0.443/0.443/0.000        0.028/0.028/0.028/0.000        6,405,249
 9     40.409%      0.534/0.535/0.535/0.000        0.027/0.027/0.027/0.000        6,358,951

 avg1  42.914%                        0.254                          0.029
 tot                                 68.650                          7.705       60,778,028

   text    data     bss     dec     hex filename
 172720    1344       8  174072   2a7f8 libz-ng.so.2
```

Levels 1-8 are a little faster, with level 6 being 1.24% faster.
Level 9 takes a small penalty due to trying really hard to find even small matches, increasing the likelihood of the matches not being matches at all.

Had to add a few more benchmark steps for this to be useful.
```
compare256/avx2/1           1.11 ns         1.11 ns   2514328460
compare256/avx2/8           1.18 ns         1.18 ns   2514327444
compare256/avx2/32          1.11 ns         1.11 ns   2514328166
compare256/avx2/96          1.95 ns         1.95 ns   1436756998
compare256/avx2/160         2.51 ns         2.51 ns   1117478410
compare256/avx2/256         3.62 ns         3.62 ns    773638652

compare256/avx512/1         1.11 ns         1.11 ns   2514328713
compare256/avx512/8         1.11 ns         1.11 ns   2514328853
compare256/avx512/32        1.11 ns         1.11 ns   2514328853
compare256/avx512/96        1.95 ns         1.95 ns   1436755135
compare256/avx512/160       2.04 ns         2.04 ns   1257161900
compare256/avx512/256       2.23 ns         2.23 ns   1257162425
```
Up to 96 byte matches they are perform practically the same.
On 160-byte matches, these results indicate it is ~19% faster.
On 256-byte matches, these results indicate it is ~39% faster.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Integrated AVX512-optimized routines to improve data comparison and matching performance on supported systems.
  
- **Tests**
  - Enhanced benchmark configurations with precise input sizes for better performance evaluation.
  - Added dedicated test cases to validate the new optimized routines for the AVX512 architecture.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->