@nmoinvaz 2025-04-10
Great idea to not use a loop like the other implementations, because the early checks always hurt performance.