@nmoinvaz 2025-04-11
Not sure if we can use qemu on x86. Grok3 says it should be supported on AVX512 with `-cpu Skylake-Server-v1`, but not sure what package on Ubuntu we would use.