@Dead2 2025-04-11
Rewrote function again to improve the speed of sub-16 byte matches by first using a 128-bit intrinsic, and after that use only 512-bit intrinsics.
This requires us to overlap on the last run, but this is cheaper than processing the tail using a 256-bit and then a 128-bit run.

Develop:
```
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     54.185%      0.078/0.078/0.079/0.000        0.030/0.031/0.031/0.000        8,526,745
 2     43.871%      0.128/0.129/0.129/0.000        0.030/0.030/0.030/0.000        6,903,702
 3     42.390%      0.158/0.158/0.158/0.000        0.029/0.029/0.029/0.000        6,670,664
 4     41.644%      0.181/0.181/0.182/0.000        0.028/0.028/0.028/0.000        6,553,205
 5     41.215%      0.196/0.197/0.197/0.000        0.028/0.028/0.028/0.000        6,485,659
 6     41.032%      0.243/0.244/0.244/0.000        0.028/0.028/0.028/0.000        6,456,912
 7     40.778%      0.337/0.338/0.338/0.000        0.028/0.028/0.028/0.000        6,416,941
 8     40.704%      0.445/0.446/0.446/0.000        0.028/0.028/0.028/0.000        6,405,249
 9     40.409%      0.528/0.528/0.529/0.000        0.027/0.027/0.027/0.000        6,358,951

 avg1  42.914%                        0.255                          0.029
 tot                                 68.956                          7.705       60,778,028

   text    data     bss     dec     hex filename
 170024    1344       8  171376   29d70 libz-ng.so.2
```

PR w/first version:
```
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     54.185%      0.079/0.079/0.079/0.000        0.031/0.031/0.031/0.000        8,526,745
 2     43.871%      0.127/0.128/0.128/0.000        0.030/0.030/0.030/0.000        6,903,702
 3     42.390%      0.156/0.156/0.157/0.000        0.029/0.029/0.029/0.000        6,670,664
 4     41.644%      0.178/0.179/0.180/0.000        0.028/0.028/0.028/0.000        6,553,205
 5     41.215%      0.193/0.194/0.194/0.000        0.028/0.028/0.028/0.000        6,485,659
 6     41.032%      0.240/0.240/0.240/0.000        0.028/0.028/0.028/0.000        6,456,912
 7     40.778%      0.335/0.335/0.335/0.000        0.028/0.028/0.028/0.000        6,416,941
 8     40.704%      0.442/0.443/0.443/0.000        0.028/0.028/0.028/0.000        6,405,249
 9     40.409%      0.534/0.535/0.535/0.000        0.027/0.027/0.027/0.000        6,358,951

 avg1  42.914%                        0.254                          0.029
 tot                                 68.650                          7.705       60,778,028

   text    data     bss     dec     hex filename
 172720    1344       8  174072   2a7f8 libz-ng.so.2
```

PR w/new version:
```
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     54.185%      0.078/0.078/0.078/0.000        0.031/0.031/0.031/0.000        8,526,745
 2     43.871%      0.127/0.127/0.128/0.000        0.030/0.030/0.030/0.000        6,903,702
 3     42.390%      0.155/0.156/0.156/0.000        0.029/0.029/0.029/0.000        6,670,664
 4     41.644%      0.179/0.179/0.180/0.000        0.028/0.028/0.028/0.000        6,553,205
 5     41.215%      0.193/0.194/0.194/0.000        0.028/0.028/0.028/0.000        6,485,659
 6     41.032%      0.239/0.240/0.240/0.000        0.028/0.028/0.028/0.000        6,456,912
 7     40.778%      0.334/0.335/0.335/0.000        0.028/0.028/0.028/0.000        6,416,941
 8     40.704%      0.442/0.442/0.442/0.000        0.028/0.028/0.028/0.000        6,405,249
 9     40.409%      0.527/0.527/0.527/0.000        0.027/0.027/0.027/0.000        6,358,951

 avg1  42.914%                        0.253                          0.029
 tot                                 68.334                          7.703       60,778,028

   text    data     bss     dec     hex filename
 172784    1344       8  174136   2a838 libz-ng.so.2
```
It is now faster on all levels except level 0 (probably too few digits in measurement).
Overall it is now 0.92% faster for all levels.
Level 2 is 1.55% faster.
Level 6 is 1.64% faster.
Even level 9 is now 0.19% faster, instead of being slower.

New comparison of AVX2 vs AVX512:
```
compare256/avx2/1           1.11 ns         1.11 ns   2514313868
compare256/avx2/10          1.11 ns         1.11 ns   2514328555
compare256/avx2/40          1.41 ns         1.41 ns   2011462645
compare256/avx2/80          1.95 ns         1.95 ns   1436753207
compare256/avx2/100         3.06 ns         3.06 ns    914300827
compare256/avx2/175         3.15 ns         3.15 ns    773638911
compare256/avx2/256         3.62 ns         3.62 ns    773638348

compare256/avx512/1        0.835 ns        0.835 ns   2514328706
compare256/avx512/10       0.835 ns        0.835 ns   3352437409
compare256/avx512/40        1.48 ns         1.48 ns   2011461596
compare256/avx512/80        1.39 ns         1.39 ns   2011225435
compare256/avx512/100       1.96 ns         1.96 ns   1430012195
compare256/avx512/175       2.66 ns         2.66 ns   1059428694
compare256/avx512/256       2.91 ns         2.91 ns    967652717
```
This is now faster on both small matches and big matches compared to the AVX2 implementation (will look into changing that too later).
1-16 byte matches is now ~25% faster than AVX2.
80-byte matches is ~29% faster
256-byte matches is ~20% faster

256-byte matches are a little slower than the 2.51ns we had with the first commit, but what we gain on the shorter matches more than makes up for it in the end.